+++ packages/sdk-ts/src/core/modules/ibc/msgs/MsgTransferCosmjs.ts
@@ -1,5 +1,11 @@
-import { MsgTransferEncodeObject } from '@cosmjs/stargate'
+// import { MsgTransferEncodeObject } from '@cosmjs/stargate'
+import snakecaseKeys, { SnakeCaseKeys } from 'snakecase-keys'
 import { MsgTransfer as BaseMsgTransferCosmjs } from 'cosmjs-types/ibc/applications/transfer/v1/tx'
+import {
+  CosmosBaseV1Beta1Coin,
+  IbcCoreClientV1Client,
+  IbcApplicationsTransferV1Tx,
+} from '@injectivelabs/core-proto-ts'
 
 export declare namespace MsgTransferCosmjs {
   export interface Params {
@@ -38,7 +44,32 @@ export default class MsgTransferCosmjs {
   }
 
   public toProto() {
-    throw new Error('Method not implemented.')
+    const { params } = this
+
+    const token = CosmosBaseV1Beta1Coin.Coin.create()
+    token.denom = params.amount.denom
+    token.amount = params.amount.amount
+
+    const message = IbcApplicationsTransferV1Tx.MsgTransfer.create()
+    message.receiver = params.receiver
+    message.sender = params.sender
+    message.sourceChannel = params.channelId
+    message.sourcePort = params.port
+    message.token = token
+
+    if (params.height) {
+      const timeoutHeight = IbcCoreClientV1Client.Height.create()
+      timeoutHeight.revisionHeight = params.height.revisionHeight.toString()
+      timeoutHeight.revisionNumber = params.height.revisionNumber.toString()
+
+      message.timeoutHeight = timeoutHeight
+    }
+
+    if (params.timeout) {
+      message.timeoutTimestamp = params.timeout.toString()
+    }
+
+    return BaseMsgTransferCosmjs.fromJSON(message)
   }
 
   public toData() {
@@ -46,34 +77,39 @@ export default class MsgTransferCosmjs {
   }
 
   public toAmino() {
-    const { params } = this
-
-    const transferMsg: MsgTransferEncodeObject = {
-      typeUrl: '/ibc.applications.transfer.v1.MsgTransfer',
-      value: BaseMsgTransferCosmjs.fromPartial({
-        sourcePort: params.port,
-        sourceChannel: params.channelId,
-        sender: params.sender,
-        receiver: params.receiver,
-        token: params.amount,
-        timeoutHeight: params.height
-          ? {
-              revisionHeight: BigInt(params.height.revisionHeight),
-              revisionNumber: BigInt(params.height.revisionNumber),
-            }
-          : undefined,
-        timeoutTimestamp: params.timeout ? BigInt(params.timeout) : undefined,
-      }),
+    const proto = this.toProto()
+    const message = {
+      ...snakecaseKeys(proto),
     }
 
-    return transferMsg
+    return {
+      type: 'cosmos-sdk/MsgTransfer',
+      value: {
+        ...message,
+      } as unknown as SnakeCaseKeys<IbcApplicationsTransferV1Tx.MsgTransfer>,
+    }
   }
 
   public toWeb3() {
-    throw new Error('Method not implemented.')
+    const amino = this.toAmino()
+    const { value } = amino
+
+    return {
+      '@type': '/ibc.applications.transfer.v1.MsgTransfer',
+      ...value,
+    }
   }
 
   public toDirectSign() {
-    throw new Error('Method not implemented.')
+    const proto = this.toProto()
+
+    return {
+      type: '/ibc.applications.transfer.v1.MsgTransfer',
+      message: proto,
+    }
+  }
+
+  public toBinary(): Uint8Array {
+    return BaseMsgTransferCosmjs.encode(this.toProto()).finish()
   }
 }